combined customer schemes
