+++ .prettierrc
@@ -1,4 +1,4 @@
-{ 
-"singleQuote": true, 
-"semi": false 
+{
+  "singleQuote": true,
+  "semi": true
 }
