+++ src/types/api.types.ts
@@ -23,3 +23,14 @@ export type sortDirection = 'asc' | 'desc'
 export type customersSortField = 'createdOn' | 'email' | 'name' | 'country'
 
 export type productsSortField = 'createdOn' | 'price' | 'name' | 'manufacturer'
+
+export interface ILoginResponseBody extends IResponseFields {
+  User: {
+    _id: string
+    username: string
+    firstName: string
+    lastName: string
+    roles: string[]
+    createdOn: string
+  }
+}
